fix: monitor scan failing due to make_interval bigint type mismatch (#20)
The idle monitor's SQL query used i64 (BIGINT) for make_interval(mins => )
but PostgreSQL's make_interval only accepts integer (i32). This caused every
scan to fail silently, preventing agents from ever being put to sleep.

Also fixed:
- RUST_LOG env missing oneclick_monitor and oneclick_scheduler crate filters
- Updated .env.example with complete logging config

+++ backend/crates/monitor/src/service.rs
@@ -10,7 +10,7 @@ use uuid::Uuid;
 
 /// How far ahead to look for upcoming scheduled jobs before deciding an agent
 /// can safely be put to sleep.
-const UPCOMING_JOB_WINDOW_MINUTES: i64 = 20;
+const UPCOMING_JOB_WINDOW_MINUTES: i32 = 20;
 
 /// Background service that detects idle agents and puts them to sleep.
 ///
@@ -135,7 +135,7 @@ impl IdleMonitor {
     /// Return all running agents that have been idle longer than
     /// `idle_timeout`.
     async fn find_idle_agents(&self) -> anyhow::Result<Vec<Agent>> {
-        let timeout_minutes = (self.idle_timeout.as_secs() / 60) as i64;
+        let timeout_minutes = (self.idle_timeout.as_secs() / 60) as i32;
 
         let agents: Vec<Agent> = sqlx::query_as(
             "SELECT * FROM agents \
